ENG-1785 added a wait command after some animations, moved command file in command folder and created new user-command file
